backend/requirements.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
boto3==1.42.63
botocore==1.42.63
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
exponent_server_sdk==2.2.0
fastapi==0.115.5
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
jmespath==1.1.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.26.0
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
python-slugify==8.0.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
text-unidecode==1.3
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.32.1
watchfiles==1.1.1
websockets==15.0.1
stripe
sentry-sdk
mixpanel
redis
gunicorn
google-genai
